requirements.txt
numpy
scipy
emcee
scikit-learn
zeus-mcmc
h5py